report-processor/requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.5
jinja2==3.1.2
boto3==1.26.137
azure-identity==1.12.0
google-cloud-storage==2.8.0
pyyaml==6.0
sqlalchemy==2.0.15
requests==2.31.0
python-dateutil==2.8.2
tabulate==0.9.0
colorama==0.4.6
click==8.1.3
